requirements.txt
streamlit==1.45.1
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.3
groq==0.13.0